launch: adt-000-on-jamendo-00
This includes some unholy fixes for very niche issues with python/magenta/pip/dill
For reasons that are unclear, it was necessary to explicitly include a few deps
in job_adt/pyproject.toml that were installed on a regular linux machine, but not
installed in the docker conteiner (or a broken version was installed).

Also there was a VERY weird issue with the `tensorflow_probability` package that
required a patch, which is currently applied in a pretty hacky way. It's worth
cleaening this up before merging back to main.